+++ backend/tests_configs.ini
@@ -302,7 +302,7 @@ user.custom_properties.ui_schema_file_path = %(user.custom_properties.dir)s/ui.j
 user.custom_properties.translations_dir_path = %(user.custom_properties.dir)s/locale
 
 [pre-filled-agenda-event_sample_test]
-app.enabled = contents/thread,contents/file,contents/html-document,contents/folder,upload_permission,share_content
+app.enabled = contents/thread,contents/file,contents/html-document,contents/folder,upload_permission,share_content,agenda
 api.key = mysuperapikey
 preview.jpg.restricted_dims = True
 email.notification.activated = false
@@ -313,7 +313,8 @@ email.notification.enabled_on_invitation = False
 webdav.ui.enabled = False
 webdav.base_url = https://localhost:3030
 webdav.root_path = /webdav
-agenda.pre_filled_event.description_file_path = %(here)s/tracim_backend/templates/pre-filled_agenda_event/test/description.txt
+caldav.radicale.storage.type = multifilesystem
+caldav.pre_filled_event.description_file_path = %(here)s/tracim_backend/templates/pre-filled_agenda_event/test/description.txt
 
 [functional_memory_storage_test]
 app.enabled = contents/thread,contents/file,contents/html-document,contents/folder,contents/kanban,upload_permission,share_content
